ai/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0

# YOLO
ultralytics>=8.0.0

# SAM (Segment Anything Model)
segment-anything>=1.0

# Model Export
onnx>=1.14.0
onnxruntime>=1.15.0
onnxscript>=0.1.0  # Required for PyTorch 2.x ONNX export
tensorflow>=2.13.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization & Explainability
matplotlib>=3.7.0
seaborn>=0.12.0
grad-cam>=1.4.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Logging & Experiment Tracking
tensorboard>=2.13.0
mlflow>=2.5.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

